% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{add.introns.feat}
\alias{add.introns.feat}
\title{Add introns to features}
\usage{
add.introns.feat(x)
}
\arguments{
\item{x}{An object of type \code{feat}.  CDS regions must be present with type "CDS", and
the transcript_id must be indicated in the attribute field.}
}
\value{
An object of type \code{feat}, with all the entries of the original object, but
also with intron annotations.
}
\description{
Add introns to features
}
\note{
If x is stored as a pointer to an object stored in C, introns will be
added to x.
}
\examples{
exampleArchive <- system.file("extdata", "examples.zip", package="rphast")
featFile <- "gencode.ENr334.gp"
unzip(exampleArchive, featFile)
f <- read.feat(featFile)
table(f$feature)
coverage.feat(f[f$feature=="CDS",])
coverage.feat(f[f$feature=="exon",])
f <- add.introns.feat(f)
table(f$feature)
coverage.feat(f[f$feature=="intron",])
unlink(featFile)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{features}

