% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{alphabet.msa}
\alias{alphabet.msa}
\title{MSA Alphabet}
\usage{
alphabet.msa(x)
}
\arguments{
\item{x}{an MSA object}
}
\value{
the valid non-missing-data characters for an MSA object.
}
\description{
Returns the alphabet used by an MSA object.
}
\examples{
m <- msa(seqs=c("a--acgtaa", "NN-nnnTAA", "AGGAGGTAG"),
         names=c("human", "mouse", "rat"))
alphabet.msa(m)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{msa}

