% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{apply.bgc.sel}
\alias{apply.bgc.sel}
\title{Apply bgc+selection parameters to a matrix}
\usage{
apply.bgc.sel(m, bgc = 0, sel = 0, alphabet = "ACGT")
}
\arguments{
\item{m}{A transition matrix}

\item{bgc}{The bgc (biased gene conversion) parameter, population-scaled.}

\item{sel}{The selection parameter (population-scaled)}

\item{alphabet}{The alphabet used for nucleotide states}
}
\value{
A matrix with bgc+sel applied.  This matrix may no longer be
time reversible.
}
\description{
Apply bgc+selection parameters to a matrix
}
\author{
Melissa J. Hubisz and Adam Siepel
}

