% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{as.list.tm}
\alias{as.list.tm}
\title{Tree Model to List}
\usage{
\method{as.list}{tm}(x, ...)
}
\arguments{
\item{x}{an object of class tm}

\item{...}{arguments to be passed to/from other functions}
}
\description{
Coerce a tree model into a list
}
\examples{
tm <- tm(tree="((human:0.01, chimp:0.01):0.03, mouse:0.3)",
         subst.mod="JC69")
is.list(tm)
is.list(as.list(tm))
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{tm}}
}

