% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{as.pointer.msa}
\alias{as.pointer.msa}
\title{MSA To Pointer}
\usage{
as.pointer.msa(src)
}
\arguments{
\item{src}{an MSA object stored by value in R}
}
\value{
an MSA object stored by reference as a pointer to an object
created in C.
}
\description{
Take an MSA stored in R and return one stored by reference
}
\examples{
m <- msa(seqs=c("A--ACGTAT", "AG-AGGTAA", "AGGAGGTAG"),
         names=c("human", "mouse", "rat"))
m
m <- as.pointer.msa(m)
m
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{msa}} for details on MSA storage options.
}
\keyword{msa}

