% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{branchlength.tree}
\alias{branchlength.tree}
\title{Get the total length of the edges of a tree}
\usage{
branchlength.tree(tree)
}
\arguments{
\item{tree}{A vector of character strings, each containing a newick tree}
}
\value{
A numeric vector containing the total branchlength of each tree
}
\description{
Get the total length of the edges of a tree
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{trees}

