% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc.R
\name{classify.muts.bgc}
\alias{classify.muts.bgc}
\title{Count the number of mutations of each gBGC type on each branch}
\usage{
classify.muts.bgc(align, mod, branch = NULL)
}
\arguments{
\item{align}{An alignment of type \code{msa}}

\item{mod}{An evolutionary model of type \code{tm}}

\item{branch}{A character vector giving the name(s) of the branch or branches we are interested in.}
}
\value{
A data frame with a row for each branch, giving the number of expected weak (A or T)
to strong (C or G) mutations, strong to weak, weak to weak, and strong to strong.  I
}
\description{
Count the number of mutations of each gBGC type on each branch
}
\author{
Melissa J. Hubisz
}

