\name{complement}
\alias{complement}
\title{complement}
\usage{complement(x)}
\description{DNA complement}
\value{The complement of the given sequence(s).  Characters other
than A,C,G,T,a,c,g,t are unchanged.}
\keyword{msa}
\arguments{\item{x}{A character vector with DNA sequences to be complemented}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
