% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{coord.range.msa}
\alias{coord.range.msa}
\title{Obtain the range of coordinates in a MSA objects}
\usage{
coord.range.msa(x, refseq = names.msa(x)[1])
}
\arguments{
\item{x}{An object of type \code{msa}}

\item{refseq}{A character string identifying the reference sequence
(or NULL to use frame of reference of entire alignment)}
}
\value{
A numeric vector of length 2 giving the smallest and highest
coordinate in the alignment.  If refseq is the first sequence in alignment,
offset.msa(x) is added to the range, otherwise it is ignored.
}
\description{
Obtain the range of coordinates in a MSA objects
}
\author{
Melissa J. Hubisz and Adam Siepel
}

