% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{flatten.feat}
\alias{flatten.feat}
\title{Combine adjacent features with the same "feature" field}
\usage{
flatten.feat(x, weightedAverageScore = FALSE, minScore = FALSE)
}
\arguments{
\item{x}{An object of type feat}

\item{weightedAverageScore}{If TRUE, scores of merged features
are given by the average of those features, weighted by their
original lengths. Otherwise, scores of merged features are
simply summed.}

\item{minScore}{If TRUE, scores of merged features are given
by the minimum score of the combined features.}
}
\value{
A features object in which adjacent features are
combined into one longer feature.
}
\description{
Combine adjacent features with the same "feature" field
}
\note{
If x is stored as a pointer to a C structure, then
x will be modified to the return value.
}
\author{
Melissa J. Hubisz and Adam Siepel
}

