% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{freq3x4.msa}
\alias{freq3x4.msa}
\title{Get codon frequencies based on 3x4 model}
\usage{
freq3x4.msa(x)
}
\arguments{
\item{x}{An object of type msa.  It is assumed to represent in-frame codons.
Length should be a multiple of 3.}
}
\value{
A vector of length 64 corresponding to the 64 codon frequencies.
The frequencies corresponding to stop codons should be 0.
}
\description{
Get codon frequencies based on 3x4 model
}
\author{
Melissa J. Hubisz
}

