\name{from.pointer.feat}
\alias{from.pointer.feat}
\title{Convert a features object from C memory (external pointer) to R memory}
\usage{
  from.pointer.feat(x)
}
\arguments{
  \item{x}{A features object stored as a pointer to C
  memory}
}
\value{
  A features object (inheriting from the data.frame class)
  stored in R memory
}
\description{
  Convert a features object from C memory (external
  pointer) to R memory
}
\author{
  Melissa J. Hubisz
}

