% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{from.pointer.msa}
\alias{from.pointer.msa}
\title{MSA From Pointer}
\usage{
from.pointer.msa(src)
}
\arguments{
\item{src}{an MSA object stored by reference}
}
\value{
an MSA object stored in R.  If src is already stored in R,
returns a copy of the object.
}
\description{
Take an MSA stored by reference and return one stored in R
}
\examples{
m <- msa(seqs=c("A--ACGTAT", "AG-AGGTAA", "AGGAGGTAG"),
         names=c("human", "mouse", "rat"), pointer.only=TRUE)
m
m <- from.pointer.msa(m)
m
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{msa}} for details on MSA storage options.
}
\keyword{msa}

