% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{gc.content.msa}
\alias{gc.content.msa}
\title{Get the fraction of G's and C's in an alignment}
\usage{
gc.content.msa(x, seq = NULL, ignore.missing = TRUE, ignore.gaps = TRUE)
}
\arguments{
\item{x}{An object of type \code{msa}}

\item{seq}{A vector of character strings identifying the sequence(s)
to use in the base frequency tabulation.  If \code{NULL}, use all
sequences.}

\item{ignore.missing}{If \code{FALSE}, count missing data in the
denominator.}

\item{ignore.gaps}{If \code{TRUE}, count gaps in the denominator.}
}
\value{
The fraction of bases which are C's and G's
}
\description{
Get the fraction of G's and C's in an alignment
}
\author{
Melissa J. Hubisz
}

