% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{hist.feat}
\alias{hist.feat}
\title{plot histogram of feature lengths}
\usage{
\method{hist}{feat}(x, type = "length", ...)
}
\arguments{
\item{x}{an object of type \code{feat}}

\item{type}{a character string, denoting the value to make the histogram with.
Currently the only valid types are "length" or "score"}

\item{...}{additional arguments to be passed to \code{hist}}
}
\description{
plot histogram of feature lengths
}
\author{
Melissa J. Hubisz
}
\keyword{features}

