% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm}
\alias{hmm}
\title{Create an rphast HMM object}
\usage{
hmm(trans.mat, eq.freq = NULL, begin.freq = NULL, end.freq = NULL)
}
\arguments{
\item{trans.mat}{A square matrix object of dimension n x n where n is the
number of states, and element [i,j] is the rate of
transition from state i to state j}

\item{eq.freq}{A vector of length n giving the equilibrium
frequencies of each state.  If NULL, calculate equilibrium frequencies
that will make a reversible markov chain.}

\item{begin.freq}{A vector of length n giving the initial state
frequencies.  If NULL, use equilibrium frequencies.}

\item{end.freq}{A vector of length n giving the final state frequencies.
If NULL, do not condition on end frequencies.}
}
\description{
Create a new HMM object
}
\examples{
h <- hmm(matrix(1, nrow=4, ncol=4))
h
}
\author{
Melissa J. Hubisz and Adam Siepel
}

