\name{is.msa}
\alias{is.msa}
\title{Check an MSA object...}
\usage{is.msa(msa)}
\description{Check an MSA object}
\value{A logical indicating whether object is of type \code{msa}}
\keyword{msa}
\arguments{\item{msa}{An object to tests}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
