% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{is.subst.mod.tm}
\alias{is.subst.mod.tm}
\title{Check Substitution Model Strings}
\usage{
is.subst.mod.tm(mod)
}
\arguments{
\item{mod}{A vector of character strings representing substitution
model names to be tested}
}
\value{
A vector of logical values indicating whether each string
represents a defined substitution model
}
\description{
Check if a string represents a phast substitution model
}
\examples{
is.subst.mod.tm(c("JC69", "K80", "F81", "HKY85", "HKY85+Gap",
                  "REV", "SSREV", "REV+GC", "UNREST", "R2", "U2", "R2S",
                  "U2S", "R3", "R3S", "U3", "U3S", "GC", "HB",
                  "bad.model"))

}
\author{
Melissa J. Hubisz and Adam Siepel
}

