\name{is.tm}
\alias{is.tm}
\title{Tree Models}
\usage{is.tm(x)}
\description{Check whether an object is of type tm (tree model)}
\value{TRUE if an object has class "tm", FALSE otherwise}
\keyword{tm}
\arguments{\item{x}{an object to be tested}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
