% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{is.track}
\alias{is.track}
\title{Is this a track?}
\usage{
is.track(x, ...)
}
\arguments{
\item{x}{An object to test}

\item{...}{ignored}
}
\value{
A logical indicating whether x is an object of type track
}
\description{
Is this a track?
}
\author{
Melissa J. Hubisz
}
\keyword{plot}

