% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{leafnames.tree}
\alias{leafnames.tree}
\title{Get the names of a tree's leaf nodes}
\usage{
leafnames.tree(object, ...)
}
\arguments{
\item{object}{A character string containing a newick tree}

\item{...}{Not currently used}
}
\value{
A character vector containing the names of the leaf nodes
}
\description{
Get the names of a tree's leaf nodes
}
\author{
Melissa J. Hubisz and Adam Siepel
}

