% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{name.ancestors}
\alias{name.ancestors}
\title{Name Ancestral Nodes}
\usage{
name.ancestors(tree)
}
\arguments{
\item{tree}{A vector of character strings, each containing a newick tree}
}
\value{
A vector of character strings containing newick trees with all
ancestors named.
}
\description{
Name ancestors of a tree
}
\examples{
trees <- c("((hg18, panTro2), mm9);",
           "((hg18:0.142679,(mm9:0.083220,rn4:0.090564):0.269385):
                0.020666,canFam2:0.193569);")
name.ancestors(trees)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{trees}

