\name{ninf.msa}
\alias{ninf.msa}
\title{The number of informative columns in an alignment...}
\usage{ninf.msa(msa)}
\description{The number of informative columns in an alignment}
\value{The number of "informative" columns in the msa.  An informative
column has at least two non-missing and non-gap characters.}
\keyword{msa}
\author{Melissa J. Hubisz and Adam Siepel}
\arguments{\item{msa}{An object of type \code{msa}}}
