% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{nrow.feat}
\alias{nrow.feat}
\title{Number of Features}
\usage{
\method{nrow}{feat}(x)
}
\arguments{
\item{x}{An object of type \code{feat}}
}
\value{
An integer containing the number of rows in each features object
}
\description{
Get the number of rows in a features object
}
\examples{
seq <- rep("hg18.chr6", 10)
src <- rep("fake_example", 10)
feature <- rep("CDS", 10)
start <- seq(1, 100, by=10)
end <- seq(10, 100, by=10)
f <- feat(seq, src, feature, start, end)
nrow.feat(f)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{features}

