% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{nstate.hmm}
\alias{nstate.hmm}
\title{HMM number of states}
\usage{
nstate.hmm(hmm)
}
\arguments{
\item{hmm}{An object of type \code{hmm}}
}
\value{
The number of states in the hidden Markov Model
}
\description{
HMM number of states
}
\author{
Melissa J. Hubisz
}
\keyword{hmm}

