\name{numleaf.tree}
\alias{numleaf.tree}
\title{Number of leaves in a Tree}
\usage{numleaf.tree(tree)
}
\description{Get the number of leaves in a tree}
\value{A numeric vector containing the number of leaves (species) in each tree}
\keyword{trees}
\author{Melissa J. Hubisz and Adam Siepel}
\arguments{\item{tree}{A vector of character strings, each containing a newick tree}
}

