% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{print.feat}
\alias{print.feat}
\title{Printing a features Object}
\usage{
\method{print}{feat}(x, ...)
}
\arguments{
\item{x}{an object of type \code{feat}}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Prints a features object.
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{write.feat}}
}
\keyword{features}

