% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc.R
\name{print.phastBiasResult}
\alias{print.phastBiasResult}
\title{Pretty-print the phastBias result list without spilling giant matrices onto the screen}
\usage{
\method{print}{phastBiasResult}(x, ...)
}
\arguments{
\item{x}{phastBias result object}

\item{...}{not used}
}
\description{
Pretty-print the phastBias result list without spilling giant matrices onto the screen
}
\author{
Melissa J. Hubisz
}

