% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{print.tm}
\alias{print.tm}
\title{Printing Tree Models}
\usage{
\method{print}{tm}(x, aslist = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{tm}.}

\item{aslist}{Logical.  If \code{TRUE}, print the tree model as a list
rather than in tree model format.}

\item{...}{arguments to be passed to/from other functions}
}
\description{
Print a tree model
}
\examples{
exampleArchive <- system.file("extdata", "examples.zip", package="rphast")
filename <- "rev.mod"
unzip(exampleArchive, filename)
tm <- read.tm(filename)
tm
print(tm, aslist=TRUE)
unlink(filename)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{tm}}
}
\keyword{tm}

