% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{rbind.feat}
\alias{rbind.feat}
\title{concatenate feature objects}
\usage{
\method{rbind}{feat}(...)
}
\arguments{
\item{...}{objects of type \code{feat} to be combined into a single object}
}
\value{
An object of type \code{feat} containing entries from all
given features
}
\description{
concatenate feature objects
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{features}

