% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{register.rphast}
\alias{register.rphast}
\title{Register RPHAST}
\usage{
register.rphast(name = "", email = "", institution = "", comments = "")
}
\arguments{
\item{name}{Your Name (Optional).  Let us know who you are, if you want.}

\item{email}{Your Email Address (Optional).  If given, we may use it
very rarely to let you know about major new releases and bug fixes.  We
will not share your email address with anyone.}

\item{institution}{Your Institution (Optional).  Let us know where
you are from.}

\item{comments}{Anything else you'd like to tell us!}
}
\description{
If you are making use of RPHAST, we would appreciate if you would let us
know.  This will send your name, email, and institution (all optional),
as well as your IP address and rphast version to our server.  We will
not share your information with anyone.
If you choose to send your email address, we may use it (very rarely) to
let you know about major new releases and bug fixes.
}
\details{
Once you register, an empty file called "registered" will be created
in ~/.rphast (non-Windows) or \%appData\%\\rphast (Windows) which will
indicate to us that you have registered, and you will no longer receive
any reminders to register when rphast is loaded.
}
\author{
Nicholas Peterson
}
\seealso{
\code{\link{nothanks.rphast}} to get rid of registration reminders
without actually registering.
}

