% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{smooth.wig}
\alias{smooth.wig}
\title{Smooth a wig plot in rphast}
\usage{
smooth.wig(coord, score, numpoints = 300)
}
\arguments{
\item{coord}{The x coordinates of un-smoothed plot}

\item{score}{The scores cooresponding to the x coordinates (should be same length as coord)}

\item{numpoints}{The number of points to use in the new plot}
}
\value{
A data frame with numpoints rows and columns "coord" and "score"
with smoothed values.  If \code{length(coord) <= numpoints}, it will contain the
original data
}
\description{
Smooth a wig plot in rphast
}
\author{
Melissa J. Hubisz
}

