% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{state.freq.msa}
\alias{state.freq.msa}
\title{Get the observed frequencies of states in an alignment}
\usage{
state.freq.msa(align, mod)
}
\arguments{
\item{align}{An object of type \code{msa}.}

\item{mod}{An object of type \code{tm} representing a tree model.}
}
\value{
A numeric vector giving the observed frequencies of each state
in the model
}
\description{
Get the observed frequencies of states in an alignment
}
\author{
Melissa J. Hubisz and Adam Siepel
}

