\name{subst.mods}
\alias{subst.mods}
\title{List PHAST Substitution Models}
\usage{
  subst.mods()
}
\value{
  a character vector with the names of all valid
  substitution models
}
\description{
  List all valid substitution models
}
\examples{
subst.mods()
}
\author{
  Melissa J. Hubisz and Adam Siepel
}

