% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{summary.feat}
\alias{summary.feat}
\title{Features Summary}
\usage{
\method{summary}{feat}(object, ...)
}
\arguments{
\item{object}{an object of type \code{feat}}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Prints a brief summary of a features object.
}
\examples{
seq <- rep("hg18.chr6", 10)
src <- rep("fake_example", 10)
feature <- rep("CDS", 10)
start <- seq(1, 100, by=10)
end <- seq(10, 100, by=10)
f <- feat(seq, src, feature, start, end)
summary(f)  # this calls summary.data.frame
summary(as.pointer.feat(f))
}
\author{
Melissa J. Hubisz
}
\keyword{features}

