\name{branchlen.tree}
\alias{branchlen.tree}
\title{Get the total length of the edges of a tree...}
\usage{branchlen.tree(tree)}
\description{Get the total length of the edges of a tree}
\value{A numeric vector containing the total branchlength of each tree}
\keyword{trees}
\arguments{\item{tree}{A vector of character strings, each containing a newick tree}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
