\name{density.feat}
\alias{density.feat}
\title{Features kernel density...}
\usage{density.feat(x, type="length", ...)}
\description{Features kernel density}
\value{A kernel density object as defined by \code{\link{density}}}
\keyword{features}
\author{Melissa J. Hubisz}
\arguments{\item{x}{An object of type \code{feat}}
\item{type}{a character string, denoting the value to compute
the density for.  Currently the only valid types are "length"
and "score"}
\item{...}{additional arguments to be passed to \code{density}}}
