% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{depth.tree}
\alias{depth.tree}
\title{Get the distance from a node to the root of a tree}
\usage{
depth.tree(tree, node)
}
\arguments{
\item{tree}{A vector of character strings, each containing a newick tree}

\item{node}{A vector of character strings, giving the node name to
use for each tree.  Will be recycled to the length of the first argument.}
}
\value{
A numeric vector containing the distance from each given
node to the root of the corresponding tree.
}
\description{
Get the distance from a node to the root of a tree
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{trees}

