\name{dim.msa}
\alias{dim.msa}
\title{Returns the dimensions of an msa object as (# of species, # of columns)...}
\usage{dim.msa(x)}
\description{Returns the dimensions of an msa object as (# of species, # of columns)}
\value{An integer vector of length two giving number of species and
number of columns in the alignment}
\keyword{msa}
\author{Melissa J. Hubisz and Adam Siepel}
\arguments{\item{x}{An object of type \code{msa}}}
