% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{enrichment.feat}
\alias{enrichment.feat}
\title{Enrichment of features with respect to annotation types}
\usage{
enrichment.feat(x, annotations, region.bounds)
}
\arguments{
\item{x}{An object of type \code{feat}}

\item{annotations}{An object of type \code{feat} containing
some annotations.}

\item{region.bounds}{An object of type \code{feat} representing
the boundary coordinates of the regions of interest (such as chromosome
boundaries).  All elements from the first two arguments should fall
entirely within region.bounds.}
}
\value{
A data frame with two columns and a row for each type of
element in the annotations.  The second column gives the 
fold-enrichment
of x across the corresponding annotation type,
which is equal to the
fraction of x which fall within the annotation type,
divided by the fraction of the entire region covered by the
annotation type.
}
\description{
Enrichment of features with respect to annotation types
}
\note{
If any of the arguments to this function are passed as pointers to
objects stored in C, they will be sorted after this function call.
}
\author{
Melissa J. Hubisz
}
\keyword{features}

