\name{hist.feat}
\alias{hist.feat}
\title{plot histogram of feature lengths...}
\usage{hist.feat(x, type="length", ...)}
\description{plot histogram of feature lengths}
\keyword{features}
\arguments{\item{x}{an object of type \code{feat}}
\item{type}{a character string, denoting the value to make the histogram with.
Currently the only valid types are "length" or "score"}
\item{...}{additional arguments to be passed to \code{hist}}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
