% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{is.msa}
\alias{is.msa}
\title{Check an MSA object}
\usage{
is.msa(msa)
}
\arguments{
\item{msa}{An object to tests}
}
\value{
A logical indicating whether object is of type \code{msa}
}
\description{
Check an MSA object
}
\author{
Melissa J. Hubisz
}
\keyword{msa}

