% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{is.tm}
\alias{is.tm}
\title{Tree Models}
\usage{
is.tm(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
TRUE if an object has class "tm", FALSE otherwise
}
\description{
Check whether an object is of type tm (tree model)
}
\author{
Melissa J. Hubisz
}
\keyword{tm}

