\name{is.track}
\alias{is.track}
\title{Is this a track?}
\usage{is.track(x, ...)}
\description{Is this a track?}
\value{A logical indicating whether x is an object of type track}
\keyword{plot}
\arguments{\item{x}{An object to test}
\item{...}{ignored}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
