% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{names.msa}
\alias{names.msa}
\title{MSA Sequence Names}
\usage{
\method{names}{msa}(x)
}
\arguments{
\item{x}{an MSA object}
}
\value{
a character vector giving the names of the sequences, or
NULL if they are not defined
}
\description{
Returns the sequence names for an MSA object.
}
\examples{
m <- msa(seqs=c("ACGTAT", "AGGTAA", "AGGTAG"),
         names=c("human", "mouse", "rat"))
names.msa(m)
m <- msa(seqs=c("ACGTAT", "AGGTAA", "AGGTAG"))
names.msa(m)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{msa}

