% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{nrow.msa}
\alias{nrow.msa}
\title{MSA Number of Sequences}
\usage{
\method{nrow}{msa}(x)
}
\arguments{
\item{x}{an MSA object}
}
\value{
an integer containing the number of sequences in an alignment.
}
\description{
Returns the number of sequence in an MSA alignment.
}
\examples{
m <- msa(seqs=c("A--ACGTAT", "AG-AGGTAA", "AGGAGGTAG"),
         names=c("human", "mouse", "rat"))
nrow.msa(m)
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\seealso{
\code{\link{msa}}
}
\keyword{msa}

