\name{nstate.hmm}
\alias{nstate.hmm}
\title{HMM number of states...}
\usage{nstate.hmm(hmm)}
\description{HMM number of states}
\value{The number of states in the hidden Markov Model}
\keyword{hmm}
\arguments{\item{hmm}{An object of type \code{hmm}}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
