% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{numleaf.tree}
\alias{numleaf.tree}
\title{Number of leaves in a Tree}
\usage{
numleaf.tree(tree)
}
\arguments{
\item{tree}{A vector of character strings, each containing a newick tree}
}
\value{
A numeric vector containing the number of leaves (species) in each tree
}
\description{
Get the number of leaves in a tree
}
\author{
Melissa J. Hubisz and Adam Siepel
}
\keyword{trees}

