\name{postprob.msa}
\alias{postprob.msa}
\title{Obtain posterior probilities of every state at every node...}
\usage{postprob.msa(x, tm)
}
\description{Obtain posterior probilities of every state at every node}
\value{An array giving the posterior probabilities of all states at
every ancestral node.}
\author{Melissa J. Hubisz and Adam Siepel}
\arguments{\item{x}{An object of type \code{msa}}
\item{tm}{An object of type \code{tm}}
}

