\name{print.feat}
\alias{print.feat}
\title{Printing a features Object}
\usage{print.feat(x, ...)}
\description{Prints a features object.}
\keyword{features}
\seealso{\code{\link{write.feat}}}
\arguments{\item{x}{an object of type \code{feat}}
\item{...}{further arguments to be passed to or from other methods}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
