\name{print.feat}
\alias{print.feat}
\title{Printing a features Object}
\usage{print.feat(x, ...)
}
\description{Prints a features object.}
\keyword{features}
\seealso{\code{\link{write.feat}}}
\author{Melissa J. Hubisz and Adam Siepel}
\arguments{\item{x}{an object of type \code{feat}}
\item{...}{further arguments to be passed to or from other methods}
}

