\name{range.feat}
\alias{range.feat}
\title{Features range}
\usage{range.feat(..., na.rm=FALSE)}
\description{Get the range of a features object}
\value{A vector of size 2 indicating minimum and maximum coord in
the features object}
\keyword{features}
\arguments{\item{...}{Objects of type \code{feat}}
\item{na.rm}{Whether to remove values of NA before calculating range.}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
