\name{range.track}
\alias{range.track}
\title{Get the coordinate range of a list of RPHAST results...}
\usage{range.track(..., na.rm=FALSE)}
\description{Get the coordinate range of a list of RPHAST results}
\value{a numeric vector of length two giving the minimum and maximum
coordinates in any element of the list}
\keyword{plot}
\arguments{\item{...}{a list of tracks}
\item{na.rm}{logical, indicating if \code{NA}'s should be omitted}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
