\name{rbind.feat}
\alias{rbind.feat}
\title{concatenate feature objects}
\usage{
  rbind.feat(...)
}
\arguments{
  \item{...}{objects of type \code{feat} to be combined
  into a single object}
}
\value{
  An object of type \code{feat} containing entries from all
  given features
}
\description{
  concatenate feature objects
}
\author{
  Melissa J. Hubisz and Adam Siepel
}
\keyword{features}

