\name{rbind.feat}
\alias{rbind.feat}
\title{concatenate feature objects...}
\usage{rbind.feat(...)}
\description{concatenate feature objects}
\value{An object of type \code{feat} containing entries from all
given features}
\keyword{features}
\arguments{\item{...}{objects of type \code{feat} to be combined into a single object}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
