\name{rename.tree}
\alias{rename.tree}
\title{Tree Node Renaming}
\usage{rename.tree(tree, old.names, new.names)}
\description{Rename nodes of trees}
\value{A vector of character strings, in which all nodes with names
given in old.names are replaced with values from new.names}
\keyword{trees}
\arguments{\item{tree}{A vector of character strings, each containing a newick tree}
\item{old.names}{A vector of current names to be substituted}
\item{new.names}{A vector of equal length to old.names giving the
substitutions}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
