\name{subst.mods}
\alias{subst.mods}
\title{List PHAST Substitution Models}
\description{List all valid substitution models}
\value{a character vector with the names of all valid substitution
models}
\examples{
subst.mods()
}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
