\name{summary.tm}
\alias{summary.tm}
\title{Tree Model Summary}
\usage{summary.tm(object, ...)}
\description{Tree model summary}
\seealso{\code{\link{tm}}}
\keyword{tm}
\arguments{\item{object}{An object of class tm}
\item{...}{Parameters to be passed to/from other functions}}
\author{Melissa Hubisz <mjhubisz@cornell.edu> and Adam Siepel
    <acs4@cornell.edu>}
